+++ specifications/src/main/resources/conformance/specifications/an/v100/changelog.md
@@ -10,6 +10,8 @@ Added `Charges` also within the `UtilizedTransportEquipment` and `CargoItem`. (S
 
 Added to `Charges` the properties `chargeCode`, `paymentLocation`, `chargePartnerCode` and `paymentCurrency`. (SD-2355, SD-2376)
 
+Added to `Location` the property `locationTypeCode`. (SD-2363)
+
 Clarified that `freightPaymentStatus` refers only to collect freight charges. (SD-2375)