+++ specifications/src/main/resources/conformance/specifications/an/v100/changelog.md
@@ -1,5 +1,26 @@
 # Arrival Notice v1.0.0 Changelog
 
+## Snapshot v1.0.0-20250815-alpha
+
+Restructured the `Transport` object to contain an array of `Leg`s, each with its own `VesselVoyage`. (SD-2352)
+
+Added classifiers (planned / estimated / actual) to `Transport` and `Leg` dates. (SD-2361, SD-2381)
+
+Added a new `ReleaseInformation` object to the root `ArrivalNotice` object. (SD-2364)
+
+Added `Charges` also within the `UtilizedTransportEquipment` and `CargoItem`. (SD-2354)
+
+Added to `Charges` the properties `chargeCode`, `paymentLocation`, `chargePartnerCode` and `paymentCurrency`. (SD-2355, SD-2376)
+
+Added property `lastFreeDate` to object `FreeTime`. (SD-2353)
+
+Added to `Location` the property `locationTypeCode`. (SD-2363)
+
+Clarified that `freightPaymentStatus` refers only to collect freight charges. (SD-2375)
+
+Clarified the semantics of the `carrierRateOfExchange` within the `Charge` object. (SD-2382)
+
+
 ## Snapshot v1.0.0-20250801-alpha
 
 Clarified that the ordering of `documentParties` is undefined except for parties of type `NI` for which it defines which one is the third, fourth, etc. notify party. (SD-2317)